+++ internal/gamestate/map.go
@@ -957,7 +957,7 @@ func (m *Map) TryMovePlayerByUserInput(sceneManager *scene.Manager, gameState *G
 	return true
 }
 
-func (m *Map) DrawFullscreenImage(screen *ebiten.Image, texture *ebiten.Image, offsetX, offsetY float64) {
+func (m *Map) DrawFullscreenImage(screen *ebiten.Image, texture *ebiten.Image, offsetX, offsetY int) {
 	op := &ebiten.DrawImageOptions{}
 	w, h := texture.Size()
 	diff := h - consts.TileYNum*consts.TileSize
@@ -981,11 +981,11 @@ func (m *Map) DrawFullscreenImage(screen *ebiten.Image, texture *ebiten.Image, o
 		op.SourceRect = &r
 	}
 
-	op.GeoM.Translate(offsetX, offsetY-float64(diff))
+	op.GeoM.Translate(float64(offsetX), float64(offsetY)-float64(diff))
 	screen.DrawImage(texture, op)
 }
 
-func (m *Map) DrawCharacters(screen *ebiten.Image, priority data.Priority, offsetX, offsetY float64) {
+func (m *Map) DrawCharacters(screen *ebiten.Image, priority data.Priority, offsetX, offsetY int) {
 	chars := []*character.Character{}
 	for _, e := range m.events {
 		page := m.currentPage(e)
